refactor(cli): split init_cmd into a structured package
Break the 2k-line init_cmd.py into a cohesive init_cmd/ package —
command, workspace, generation, persistence, reporting, and the
interactive prompts — and collapse the duplicated single-repo and
per-repo workspace generation paths into one shared core (coverage
selection, cost gate, generation + KG enrichment + cost flush).

Also replace the two hand-rolled config.yaml yaml.dump writes with
save_config_partial (extended to take arbitrary keys), and route the
DB-engine setup through the existing open_repo_db helper.

The public import surface is preserved: init_command plus the
leading-underscore names sibling commands (search/reindex) and the
test-suite import from repowise.cli.commands.init_cmd still resolve.
